third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-panel-rewrite-href.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<link rel="stylesheet" src="/bogusSheet1.css">
<link rel="stylesheet" src="bogusSheet2.css">
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.expandElementsTree(step1);

  function step1() {
    var innerMapping = TestRunner.domModel._idToDOMNode;

    var outputLines = [];
    for (var nodeId in innerMapping) {
      var node = innerMapping[nodeId];
      if (node.nodeName() === 'LINK' || node.nodeName() === 'SCRIPT') {
        var segments = [];
        var href = node.resolveURL(node.getAttribute('src') || node.getAttribute('href'));
        if (!href) {
          segments.push('<empty>');
          continue;
        }
        if (href.startsWith('http:')) {
          outputLines.push(href);
          continue;
        }
        var parsedURL = href.asParsedURL();
        if (!parsedURL) {
          testController.notifyDone('FAIL: no URL match for <' + href + '>');
          continue;
        }
        var split = parsedURL.path.split('/');
        for (var i = split.length - 1, j = 0; j < 3 && i >= 0; --i, ++j)
          segments.push(split[i]);
        outputLines.push(segments.reverse());
      }
      if (node.nodeName() === 'A')
        outputLines.push(node.resolveURL(node.getAttribute('href')));
    }
    outputLines.sort();
    TestRunner.addResult(outputLines.join('\n'));
    TestRunner.completeTest();
  }
}

</script>
</head>

<body>
<p>
Tests that src and href element targets are rewritten properly.
</p>

<a style="display:none" href=" javascript:alert('foo') "></a>
<iframe src="resources/elements-panel-rewrite-href-iframe.html" onload="runTest()"></iframe>
</body>
</html>
