third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-panel-rewrite-href-expected.txt

Tests that src and href element targets are rewritten properly.


 javascript:alert('foo') 
http://127.0.0.1/script.js
http://127.0.0.1/stylesheet.css
http://127.0.0.1/target.html
http://127.0.0.1:8000/bogusSheet1.css
http://127.0.0.1:8000/devtools/elements/bogusSheet2.css
http://127.0.0.1:8000/inspector/elements-test.js
http://127.0.0.1:8000/inspector/inspector-test.js
